requirements.txt
Django==4.0.10
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
pymongo==3.12.2

